Phase 4: hand-port asset-loader + pointer-arg scaling for params
The binary's master asset-load loop FUN_0042ef8a never completed because
of three broken backward gotos that each early-returned after one
iteration of the file-table walk. With it returning prematurely, no
file handles ever opened and the asset table at 0x009b22f0 stayed empty.

Hand-ported FUN_0042ef8a (decompiled/c/42ef8a.c) as a structured
labelled-loop, restoring the full iterate-22-files behaviour. Also:

- FUN_0042f239: hand-port the asset-path builder; reads file index
  from regs.ebx (set by the asset-loader caller before each iteration)
  and concatenates base path + filename pointer.
- FUN_004083b5: hand-port the file-open wrapper; treats small param_1
  as an asset-table index (binary's internal convention, not the C
  signature Ghidra emitted).

Translator change: pointer-arithmetic scaling now applies to function
parameters, not just locals. Without this, `uint *param_2; param_2 =
param_2 + 1;` in C strcpy-style functions stride 1 byte instead of 4,
turning dword-scan loops into infinite re-reads. Affected ~80 files.

Tests: 116/116 still pass. Translator: 1214/1214 still generate.

Asset-loader now runs but execution hits a new infinite loop downstream
(209M ops in 30s); narrowing under investigation.

+++ ported/auto/413800.js
@@ -17,7 +17,7 @@ export function FUN_00413800(heap, _Str, _Ch) {
     }
     iVar2 = ((iVar2 + -1) >>> 0);
     pcVar4 = ((_Str + 1) >>> 0);
-    cVar1 = ((heap.u32(_Str)) & 0xff);
+    cVar1 = ((heap.i8(_Str)) & 0xff);
     _Str = ((pcVar4) >>> 0);
   } while (cVar1 != 0);
   iVar2 = ((-(iVar2 + 1)) >>> 0);
